refactored App.js into Home.js
